+++ src/reducers/index.js
@@ -1,7 +1,7 @@
 import * as types from '../constants/action_types.js';
-import deepAssign from 'deep-assign';
+import merge from 'merge-options';
 
-const initialState = {
+const getInitialState = () => ({
   // Options set on initialization
   api: 'https://api.mapbox.com/directions/v5/',
   profile: 'mapbox/driving-traffic',
@@ -22,6 +22,10 @@ const initialState = {
     profileSwitcher: true,
     inputs: true,
     instructions: true
+  },
+
+  instructions: {
+    showWaypointInstructions: true
   },
 
   // Optional setting to pass options available to mapbox-gl-geocoder
@@ -46,14 +50,16 @@ const initialState = {
 
   // Directions data
   directions: [],
+  fetchDirectionsRequest: null,
   routeIndex: 0, 
   routePadding: 80
-};
+});
 
-function data(state = initialState, action) {
+function data(state = getInitialState(), action) {
   switch (action.type) {
-  case types.SET_OPTIONS:
-    return deepAssign({}, state, action.options);
+  case types.SET_OPTIONS: {
+    return merge({}, state, action.options);
+  }
 
   case types.DIRECTIONS_PROFILE:
     return Object.assign({}, state, {
@@ -117,10 +123,15 @@ function data(state = initialState, action) {
       waypoints: [],
       directions: []
     });
-
+  
+  case types.DIRECTIONS_REQUEST_START:
+    return Object.assign({}, state, {
+      fetchDirectionsRequest: action.request
+    });
   case types.DIRECTIONS:
     return Object.assign({}, state, {
-      directions: action.directions
+      directions: action.directions,
+      fetchDirectionsRequest: null
     });
 
   case types.ROUTE_INDEX: